requirements.txt
streamlit
pandas
python-dotenv
snowflake-snowpark-python
snowflake-connector-python
snowflake-ml-python
snowflake.core